requirements.txt
pywin32>=306
python-dotenv>=1.0.0
pyserial>=3.5
